extracted logic from ReaderPostService into /reader/service/post/ReaderPostLogic
